[topo] Id & Callsite to their own module.
